ml_service/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
lightgbm
psycopg2-binary
sqlalchemy
pydantic
joblib
shap
python-dotenv
requests
